+++ R/fastglmm.nb.R
@@ -20,22 +20,18 @@
 #'
 #' @examples
 #' library(MASS)
-#' library(lme4)
-#' 
-#' set.seed(101)
-#' dd <- expand.grid(f1 = factor(1:3),
-#'                f2 = LETTERS[1:2], g=factor(1:9), rep=1:15,
-#'        KEEP.OUT.ATTRS=FALSE)
-#' mu <- 5*(-4 + with(dd, as.integer(f1) + 4*as.numeric(f2)))
-#' dd$y <- rnbinom(nrow(dd), mu = mu, size = 0.5)
-#' 
-#' # NB GLMM via Laplace approximation
-#' fit1 <- glmer.nb(y ~ f1*f2 + (1|g), data=dd)
+#' data(PsychAD)
+#'
+#' # regression formula
+#' form <- PTPRG ~ (1|SubID) + offset(log(libSize))
+#'
+#' # NB GLMM on PTPRG expression via PQL
+#' fit1 <- fastglmm.nb(form, PsychAD)
 #' coef(summary(fit1))
 #' 
-#' # NB GLMM via PQL
-#' fit2 <- fastglmm.nb(y ~ f1*f2 + (1|g), data=dd)
-#' coef(summary(fit2))
+#' # NB GLMM via Laplace approximation
+#' # fit2 <- lme4::glmer.nb(form, PsychAD)
+#' # coef(summary(fit2))
 #
 #' @importFrom MASS negative.binomial
 #' @export